kód modularizálása, néhány módosítás, pár funkció kiszervezése osztályba
