+++ app/src/main/java/aa/droid/norbo/projects/edzesnaplo3/mvvm/ui/TevekenysegActivity.java
@@ -49,6 +49,7 @@ import aa.droid.norbo.projects.edzesnaplo3.mvvm.ui.rcviews.KorabbiSorozatRcViewA
 import aa.droid.norbo.projects.edzesnaplo3.mvvm.ui.utils.DateTimeFormatter;
 import aa.droid.norbo.projects.edzesnaplo3.mvvm.ui.utils.NaploListFactory;
 import aa.droid.norbo.projects.edzesnaplo3.mvvm.ui.utils.naplo.NaploWorker;
+import aa.droid.norbo.projects.edzesnaplo3.mvvm.ui.utils.naplo.SorozatUtil;
 import aa.droid.norbo.projects.edzesnaplo3.mvvm.ui.viewmodels.SorozatViewModel;
 import dagger.hilt.android.AndroidEntryPoint;
 
@@ -68,6 +69,9 @@ public class TevekenysegActivity extends BaseActiviry<MvvmActivityTestBinding> i
 
     @Inject
     EdzesTervValasztoDialog tervValasztoDialog;
+
+    @Inject
+    SorozatUtil sorozatUtil;
 
     @Inject
     SharedPreferences sharedPreferences;
@@ -188,77 +192,14 @@ public class TevekenysegActivity extends BaseActiviry<MvvmActivityTestBinding> i
 
         sorozatViewModel.getSorozatByGyakorlat(gyakorlatId).observe(this, sorozats -> {
             if(sorozats != null && sorozats.size() > 0) {
-                ArrayAdapter<NaploEsSorozat> listAdapter = new ArrayAdapter<NaploEsSorozat>(this, R.layout.mvvm_korabbi_sorozat_item, makeNaploEsSorozat(sorozats)) {
-                    MvvmKorabbiSorozatItemBinding itemBinding;
-                    @NonNull
-                    @Override
-                    public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
-                        if(convertView == null) {
-                            convertView = LayoutInflater.from(TevekenysegActivity.this).inflate(R.layout.mvvm_korabbi_sorozat_item, parent, false);
-                            itemBinding = DataBindingUtil.bind(convertView);
-                            convertView.setTag(itemBinding);
-                        } else {
-                            itemBinding = (MvvmKorabbiSorozatItemBinding) convertView.getTag();
-                        }
-
-                        NaploEsSorozat item = getItem(position);
-                        itemBinding.korabbiSorozatDatumLabel.setText(formatter.getNaploDatum(item.naplodatum)+" "+getSorozatOsszSuly(item.sorozats));
-                        itemBinding.korabbiSorozatLista.setAdapter(new KorabbiSorozatRcViewAdapter(item.sorozats, formatter));
-                        itemBinding.korabbiSorozatLista.setLayoutManager(new LinearLayoutManager(TevekenysegActivity.this, RecyclerView.HORIZONTAL, false));
-                        return itemBinding.getRoot();
-                    }
-                };
-
                 new AlertDialog.Builder(this)
                         .setTitle("Sorozatok")
-                        .setAdapter(listAdapter, null)
+                        .setAdapter(sorozatUtil.getSorozatEsNaploAdapter(sorozats), null)
                         .setPositiveButton("ok", (dialog, which) -> dialog.dismiss())
                         .show();
             } else {
                 Toast.makeText(this, "Nincs rögzítve még sorozat evvel a gyakorlattal", Toast.LENGTH_SHORT).show();
             }
         });
-    }
-
-    private List<NaploEsSorozat> makeNaploEsSorozat(List<Sorozat> sorozats) {
-        List<NaploEsSorozat> list = new ArrayList<>();
-        Set<Long> naploDatumok = sorozats.stream().map(Sorozat::getNaplodatum).collect(Collectors.toSet());
-        naploDatumok.forEach(aLong -> {
-            NaploEsSorozat naploEsSorozat = new NaploEsSorozat(aLong);
-            List<Sorozat> sorozats1 = new ArrayList<>();
-            sorozats.forEach(sorozat -> {
-                if(aLong == (sorozat.getNaplodatum()))
-                    sorozats1.add(sorozat);
-            });
-            naploEsSorozat.setSorozats(sorozats1);
-            list.add(naploEsSorozat);
-        });
-        list.sort((o1, o2) -> Long.compare(o2.naplodatum, o1.naplodatum));
-        return list;
-    }
-
-    private String getSorozatOsszSuly(List<Sorozat> sorozats) {
-        return String.format(Locale.getDefault(), "%,d Kg", sorozats.stream().mapToInt(sor -> sor.getSuly() * sor.getIsmetles()).sum());
-    }
-
-    public class NaploEsSorozat {
-        private long naplodatum;
-        private List<Sorozat> sorozats;
-
-        public NaploEsSorozat(long naplodatum) {
-            this.naplodatum = naplodatum;
-        }
-
-        public void setSorozats(List<Sorozat> sorozats) {
-            this.sorozats = sorozats;
-        }
-
-        public List<Sorozat> getSorozats() {
-            return sorozats;
-        }
-
-        public long getNaplodatum() {
-            return naplodatum;
-        }
     }
 }
